+++ retrorecon/dynamic_schemas.py
@@ -20,70 +20,6 @@ def register_demo_schemas(registry: SchemaRegistry) -> None:
     )
 
 
-    registry.register(
-        "subdomonster_page",
-        {
-            "required": ["init_script"],
-            "content": [
-                {
-                    "tag": "div",
-                    "attrs": {"id": "subdomonster-overlay", "class": "notes-overlay hidden"},
-                    "children": [
-                        {
-                            "tag": "div",
-                            "attrs": {"class": "mb-05 search-input-row"},
-                            "children": [
-                                {
-                                    "tag": "button",
-                                    "attrs": {"type": "button", "class": "btn overlay-close-btn", "id": "subdomonster-close-btn"},
-                                    "text": "Close",
-                                },
-                                {
-                                    "tag": "input",
-                                    "attrs": {
-                                        "type": "text",
-                                        "id": "subdomonster-domain",
-                                        "class": "form-input mr-05 subdomonster-domain-bar",
-                                        "placeholder": "example.com",
-                                    },
-                                },
-                                {
-                                    "tag": "select",
-                                    "attrs": {"id": "subdom-source", "class": "form-select mr-05"},
-                                    "children": [
-                                        {"tag": "option", "attrs": {"value": "crtsh"}, "text": "crt.sh"},
-                                        {"tag": "option", "attrs": {"value": "virustotal"}, "text": "VirusTotal"},
-                                        {"tag": "option", "attrs": {"value": "local"}, "text": "Local"},
-                                    ],
-                                },
-                                {
-                                    "tag": "input",
-                                    "attrs": {
-                                        "type": "text",
-                                        "id": "subdom-api",
-                                        "class": "form-input mr-05 w-10em",
-                                        "placeholder": "API key",
-                                    },
-                                },
-                                {
-                                    "tag": "button",
-                                    "attrs": {"type": "button", "class": "btn", "id": "subdom-fetch-btn"},
-                                    "text": "Import",
-                                },
-                                {
-                                    "tag": "span",
-                                    "attrs": {"id": "subdomonster-status", "class": "ml-05"},
-                                },
-                            ],
-                        },
-                        {"tag": "div", "attrs": {"id": "subdomonster-table", "class": "mt-05"}},
-                        {"tag": "div", "attrs": {"id": "subdomonster-pagination", "class": "mt-05"}},
-                        {"html_field": "init_script"},
-                    ],
-                }
-            ],
-        },
-    )
 
     registry.register(
         "screenshotter_page",
