+++ static/domain_summary.js
@@ -7,6 +7,42 @@ function initDomainSummary(){
   const totalHosts = document.getElementById('summary-total-hosts');
   const topList = document.getElementById('summary-top-list');
   const lonelyList = document.getElementById('summary-lonely-list');
+  const domainInput = document.getElementById('subdomain-import-domain');
+  const sourceSel = document.getElementById('subdom-source');
+  const apiInput = document.getElementById('subdom-api');
+  const fetchBtn = document.getElementById('subdom-fetch-btn');
+  const statusSpan = document.getElementById('subdom-status');
+
+  let statusTimer = null;
+  let statusDelay = 1000;
+
+  function showStatus(msg){
+    if(window.showStatus){
+      window.showStatus(msg);
+    } else if(statusSpan){
+      statusSpan.textContent = msg;
+      setTimeout(() => { if(statusSpan.textContent === msg) statusSpan.textContent = ''; }, 4000);
+    }
+  }
+
+  function pollStatus(){
+    if(overlay.classList.contains('hidden')){ statusTimer = null; return; }
+    fetch('/status')
+      .then(r => r.status === 204 ? null : r.json())
+      .then(data => {
+        if(data && data.code && data.code.startsWith('subdomonster')){
+          showStatus(data.message || data.code);
+          statusDelay = 1000;
+        } else {
+          statusDelay = Math.min(statusDelay * 2, 3000);
+        }
+        statusTimer = setTimeout(pollStatus, statusDelay);
+      })
+      .catch(() => { statusTimer = setTimeout(pollStatus, 5000); });
+  }
+
+  function startStatusPolling(){ if(!statusTimer) pollStatus(); }
+  function stopStatusPolling(){ if(statusTimer){ clearTimeout(statusTimer); statusTimer = null; } }
 
   async function loadSummary(){
     try{
@@ -33,6 +69,7 @@ function initDomainSummary(){
   closeBtn.addEventListener('click', () => {
     overlay.classList.add('hidden');
     document.body.style.overflow = '';
+    stopStatusPolling();
     if(location.pathname === '/tools/domain_summary'){
       history.pushState({}, '', '/');
     }
@@ -43,6 +80,28 @@ function initDomainSummary(){
       closeBtn.click();
     }
   });
+
+  if(fetchBtn){
+    fetchBtn.addEventListener('click', async () => {
+      const domain = domainInput ? domainInput.value.trim() : '';
+      const source = sourceSel ? sourceSel.value : 'crtsh';
+      if(source !== 'local' && !domain) return;
+      const params = new URLSearchParams({source});
+      if(domain) params.set('domain', domain);
+      if(apiInput && apiInput.value.trim()) params.set('api_key', apiInput.value.trim());
+      startStatusPolling();
+      try{
+        const resp = await fetch('/subdomains', {method:'POST', headers:{'Content-Type':'application/x-www-form-urlencoded'}, body:params});
+        if(resp.ok){
+          await loadSummary();
+        } else {
+          showStatus(await resp.text());
+        }
+      }catch(err){
+        console.error('import failed', err);
+      }
+    });
+  }
 
   loadSummary();
   setInterval(loadSummary, 5000);
