+++ templates/subdomain_summary.html
@@ -1,6 +1,17 @@
 <div id="domain-summary-overlay" class="notes-overlay hidden">
   <button type="button" class="btn overlay-close-btn" id="domain-summary-close-btn">Close</button>
   <h1>Subdomain Summary</h1>
+  <div class="mb-05 search-input-row">
+    <input type="text" id="subdomain-import-domain" class="form-input mr-05 subdomonster-domain-bar" placeholder="example.com" />
+    <select id="subdom-source" class="form-select mr-05">
+      <option value="crtsh">crt.sh</option>
+      <option value="virustotal">VirusTotal</option>
+      <option value="local">Local</option>
+    </select>
+    <input type="text" id="subdom-api" class="form-input mr-05 w-10em" placeholder="API key" />
+    <button type="button" class="btn" id="subdom-fetch-btn">Import</button>
+    <span id="subdom-status" class="ml-05"></span>
+  </div>
   <p>Total root domains: <span id="summary-total-domains">{{ total_domains }}</span></p>
   <p>Total unique hosts: <span id="summary-total-hosts">{{ total_hosts }}</span></p>
   <h2>Top 5 Subdomains</h2>
